Rename cells_table file to match Cells struct
