+++ auraed/src/runtime/cells/mod.rs
@@ -1,7 +1,7 @@
 use cell::Cell;
 use cell_name::CellName;
 pub(crate) use cell_service::CellService;
-use cells_table::Cells;
+use cells::Cells;
 use error::{CellsError, Result};
 use executable::Executable;
 use executable_name::ExecutableName;
@@ -9,7 +9,8 @@ use executable_name::ExecutableName;
 mod cell;
 mod cell_name;
 mod cell_service;
-mod cells_table;
+#[allow(clippy::module_inception)]
+mod cells;
 mod error;
 mod executable;
 mod executable_name;
